Improvement to VIF
Based on using the correlation matrix to calculate VIF, it is not possible to calculate VIF for terms that are always a constant. Added a check to  exclude constant terms from the VIF calculations. This prevents errors from occuring if a global intercept term is included or if all datapoints are in the same series